Fixed warning in Renderer2D Warning in Renderer2D::EndScene() of implicit conversion from ptrdiff_t to uint32_t

+++ GameEngine/src/Engine/Renderer/Renderer2D.cpp
@@ -125,7 +125,7 @@ namespace engine
 	{
 		ENGINE_PROFILE_FUNCTION();
 
-		uint32_t dataSize = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)(&s_Data.QuadVertexBufferBase[0]);
+		uint32_t dataSize = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)&s_Data.QuadVertexBufferBase[0]);
 		s_Data.QuadVertexBuffer->SetData(&s_Data.QuadVertexBufferBase[0], dataSize);
 		
 		Flush();
